update MultiObjectiveOptimizer and objective_functions
